django_testin_app/coronavstech/companies/views.py
from django.shortcuts import render
from rest_framework.viewsets import ModelViewSet
from rest_framework.pagination import PageNumberPagination

from django_testin_app.coronavstech.companies.models import Company
from django_testin_app.coronavstech.companies.serializer import CompanySerializer

class CompanyViewSet(ModelViewSet):
    serializer_class = CompanySerializer
    queryset = Company.objects.all().order_by("-last_update")
    pagination_class = PageNumberPagination
