pyproject.toml
[tool.poetry]
name = "django_testin_app"
version = "0.1.0"
description = ""
authors = ["janislaus <jankaltenbacher@gmail.com>"]
packages =[{include = "django_testin_app"}]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"
black = "^21.6b0"
pytest-django = "^4.4.0"
pytest-pythonpath = "^0.7.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
